Edit to furniture in map prefab, edit to pathfinding prefab
READ THIS BEFORE PROCEEDING: i removed everything from scene to avoid conflicts. Everything that is required to be in the scene is now to be placed in the ASSETS>RESOURCES Folder. PLEASE NOTE: if an NPC is required in the scene (Mr. Gerald or Bug) and not dynamically spawned through code, its FIRST TARGET must be define in its PATHFINDING SCRIPT!!!!. The PATHFINDING PREFAB requires you to add the TRIANGULAR NODE PREFAB FOR THE HIERARCHY.       IMPORTANT: #1 SAVE EVERYTHING YOU DO AS A PREFAB UNDER THAT RESOURCE FOLDER AND DELETE IT FROM THE SCENE BEFORE COMMITTING AND SENDING OF TO GIT. ADD EVERYTHING BACK WHEN YOU EDIT AND MAKE A NOTE ON SLACK AS TO WHAT NEEDS TO BE DONE FOR THAT PREFAB TO WORK IF ANYTHING. #2 WHEN EDITING DON'T DRAG DIRECTLY INTO THE SCENE, BUT DRAG THE PREFAB TO THE HIERARCHY TAB TO PRESERVE LOCATIONS OF THE PREFABS.